Refactor AudioMixer
